chore(deps): break ArmorPipLayout 3-way cycle
Move `Bounds` class from `ArmorPipLayout.ts` to `ArmorPipLayout.types.ts`
so the types module becomes a true leaf with zero sibling imports. The
orchestrator re-exports `Bounds` for backward compatibility (no external
consumer breakage), and `.grouped.ts` / `.processing.ts` now import
`Bounds` directly from `.types.ts` instead of routing through `.ts`.

Closes 3 circular dependencies inside src/services/printing/:
  1) ArmorPipLayout.grouped.ts  > ArmorPipLayout.ts
  2) ArmorPipLayout.ts          > ArmorPipLayout.processing.ts
  3) ArmorPipLayout.ts          > ArmorPipLayout.processing.ts > ArmorPipLayout.types.ts

After: `npx madge --circular --extensions ts,tsx src/services/printing/`
reports zero cycles. Zero behavior change — public API preserved via
re-export. All 285 printing tests pass.
